requirements.txt
langchain-google-genai 
langchain
python-dotenv
langchain_community
google-generativeai
langchain_core
streamlit

